refactor(schema): share the usage-pricing arithmetic
`costOf(price, usage)` moves into `model/pricing.ts` so the cost model and
any other caller derive a figure the same way, rather than each repeating the
per-column sum. A nullable column is an unread rate: tokens billed against one
make the whole call unknown instead of silently undercounting it.
